app/backend/requirements.txt
openai
python-dotenv
fastapi
uvicorn[standard]
odmantic
jwt
python-multipart